+++ src/tools/git/commitlint.js
