+++ src/tools/js/prettier.js
@@ -3,9 +3,12 @@ const prettier = require('prettier')
 
 const log = require('@dhis2/cli-helpers-engine').reporter
 
-const { readFile, writeFile } = require('./files.js')
+const { readFile, writeFile } = require('../../files.js')
 
-const prettierConfig = path.join(__dirname, '../config/prettier.config.js')
+const prettierConfig = path.join(
+    __dirname,
+    '../../../config/prettier.config.js'
+)
 log.debug('Prettier configuration file', prettierConfig)
 
 /**
